requirements.txt
discord.py
google-genai
python-dotenv
PyPDF2
requests
beautifulsoup4
pillow
youtube-transcript-api
aiohttp